Implement engine disabling
- base class Engine - disabled engine
- derived class EngineImpl - active engine
Configuration parameter disabled determines which one is used
